sikayetdurum int -> string

+++ SehirAsistanim.Infrastructure/Services/SikayetService.cs
@@ -42,7 +42,7 @@ namespace SehirAsistanim.Infrastructure.Services
                 FotoUrl = s.FotoUrl,
                 GonderilmeTarihi = s.GonderilmeTarihi,
                 CozulmeTarihi = s.CozulmeTarihi,
-                Durum = (int)s.Durum,
+                Durum = s.Durum,
                 DogrulanmaSayisi = s.DogrulanmaSayisi,
                 Silindimi = s.Silindimi,
